+++ SpaghettiEngine/src/main/java/com/spaghetti/interfaces/Renderable.java
@@ -1,9 +1,9 @@
 package com.spaghetti.interfaces;
 
-import org.joml.Matrix4d;
+import org.joml.Matrix4f;
 
 public interface Renderable {
 
-	public abstract void render(Matrix4d projection, double delta);
+	public abstract void render(Matrix4f projection, float delta);
 
 }
